[connman] Include usage from unavailable network services.
When a counter is first registered immediately send it the usage stats
of all known network services. This includes both currently available
and unavailable network services. Excludes network services of type
cellular as it is desirable to only show statistics for the currently
inserted SIM.

A new ResetCounters method is added to allow resetting all statistics
for the specified type of network service.