[connman] Change data counters to use 64 bit data.
The previous data counters used unsigned 32 bit types. This is
probably adequate for most of the data fields, but for the bytes
received and bytes transmitted it means that the counters wrap after
4GB of data.

Change to use the 64 bit interface to retrieve the data from the kernel
and to use 64 bit types in both the storage and DBus API.

This changes the type signature of the DBus interface and expects
applications to cope.

This changes the binary format of the data files, record size are now
larger. The magic number is changed which causes the data files to be
reinitialised losing previous data history. But that is ok because it
is likely to have wrapped anyway.