[connman] IP address collision fix
Fixes a problem with tethering switching off when MMS context is activated.
Or at least one of the scenarios when it might happen.

This is actually a merge of this commit from kernel.org:

http://git.kernel.org/cgit/network/connman/connman.git/commit/?id=60ee35f0

The IP address collision was incorrectly checked. The check reported
collisions even if there was none. This had the effect that if the
tethering uplink connection went down and then up, the tethering address
was released and then new address was allocated without a reason.