[connman] Return an in progress error if the service is already connecting.
If the network service is already connecting return an in progress
error and do nothing. Previously only the pending variable was tested
indicating that a previous DBus call started the connection. This
variable is not set for an internally initiated connection.

If the connect_sevice call is not aborted it is possible that the
service connection will be needlessly aborted when looking for an
available interface.