[connman] check service ID validity
storage: check service ID validity

Connman service ID is used, in the form it is read from storage, as
part of D-Bus path in connmand's D-Bus API. If the path constructed
using the service ID does not conform to D-Bus specifications, D-Bus
library assertion will cause connmand to exit; on restart the same
invalid ID is read from the storage and the assertion is triggered
again.

I experienced a connmand restart loop because some apparently
wispr-related memory corruption caused a service ID with binary
characters to be stored on disk. While that bug should be fixed also,
sanity-checking the service ID does not harm.