ipconfig: No need to remove the gateway when adding new one
If we remove the gateway when adding new one, then that will
prevent the VPN restart from working in all network configurations.
The reason is that if we remove the gateway here, then there will
not be any active gateway when the VPN is re-activated. This then
means that in connection.c:set_vpn_routes() the function will not
be able set the VPN server routes because the active_gateway variable
will be NULL as we just removed it. So simple solution is to not
remove the gateway.