peer: Provide the proper ipv4 settings when peer is the dhcp server
From core point of view, getting the IPv4 settings is different wheather
the peer is running dhcp as a client or as a server. On client mode it
can put those settings into DBus from its ipconfig structure. On server
mode, it has to get the details from the dhcp server structure and build
the DBus content by itself. In the end this does not change anything for
the end-user.