service: Fix service disconnection when using the same index
When a service is to be connected, other service(s) using the same
interface will be disconnected. Fix the already connected or connecting
situation by ignoring the service in question when seen in the list.

Simplify the index detection and signal a timeout other services are
disconnected in order to indicate to the caller that the connect
action needs to be redone.

Thanks to Chengyi Zhao for finding the issue.

Conflicts:
	connman/src/service.c