service: Reply to disconnection from association and configuration states
A D-Bus reply to the client calling Connect() has been properly sent
when entering ready and failure states and a also on a proper disconnect
when calling __connman_service_disconnect().

But when a service enters disconnect state from either association or
configuration states, sending of the D-Bus reply has not happened.