gsupplicant: Call p2p support callback in relevant places
At the time we detect p2p support, interface might not be related to any
wifi plugin structure. While this was not a problem before, it becomes
one now that peer_service.c is in use: as soon as the peer driver is
set, it will try to register all existing peer services to the
interfaces, and without any relation between wifi<->interface, this was
generating a crash bug.

The issue is now fixed as the p2p support callback is called after
interface_create_callback(), which sets the relation between wifi and
interface.