[connman] Re-calculate DHCP timers when waking up from suspend
Modify jolla_wakeup_timer so that it keeps track of the wakeup timers,
executes them when they expire and re-calculates wait timeout when
coming out of suspend.

This is to enable reasonable DHCP timeout performance even when using
CLOCK_MONOTONIC, as converting glib to use CLOCK_BOOTTIME proved to be
more difficult than expected.