config: Refactor connecting after service provisioning
Refactor the code that implicitely called __connman_service_connect()
for ethernet and gadget services and __connman_service_autoconnect()
for a non-NFC provisioned wifi service to do this explicitely.

Ethernet and gadget services are to be connected immediately when
provisioned as the services only exist while a cable is attached. For
a non-NFC provisioned wifi service autoconnect is used, as the
provisioned wifi might not be the best one in range. An NFC
provisioned wifi, i.e. the only category that currently has a virtual
config,  will be connected explicitely after the exiting to the main
loop.