network: Set IPv6 ipconfig if it is missing
Following scenario is possible:
- we have a wifi service A connected and IPv6 is enabled for it
- we connect to other IPv6 enabled service B
- the first service is disconnected
- kernel sends rtnl messages related to events that are done   when
  we connect/disconnect the services
- if A service's link down event is received after we have received
  B's link up, then we will remove the ipv6 config from service B
- this has the effect that IPv6 addresses are not known by service B
  any more

So when we receive a reply to router solicitation message, we check
if the IPv6 config is still there in service B. If it is not, then
we can create it so that service gets the IPv6 config back.

The missing IPv6 config can only happen if we are switching from one
service to another. If we disconnect first service A and then
connect service B, no issues were seen.