[connman] Fix D-Bus call cancellation for interface creation and removal
Property-retrieving D-Bus calls made within supplicant interface
creation sequence must be tagged with null caller pointer, as wifi
plugin which may invoke cancellation does not yet know the interface
pointer value.

D-Bus call for interface removal must be tagged with the interface
structure pointer so that it can be cancelled if the interface is
otherwise removed during the time the call is pending.