service: Do not reply via D-Bus when clearing service error
When clearing a service in error state, set both ipconfigs to idle
causing the service states in both ipconfig and service to be reset.
Save the D-Bus id before setting the service to idle and restore it
afterwards so that a possible D-Bus client won't be informed of a
connection error which actually never happened.

Clean up unnecessary checks at the same time.