[connman] Give some time for bluetoothd to start up
In rare cases, connmand would rfkill BT at a time when bluetoothd was
still in the middle of initializing the device, and as a result
bluetoothd might end up in a state from which it would not recover
successfully.

Work around the issue by delaying rfkilling BT slightly after a BT
adapter appears on D-Bus.