[connman] storage: Settings file read caching
Implemented a read cache for settings files, so that files are not needlessly
read and parsed every time some configuration values is needed. To keep things
simple, any modification of the file on disk will drop the corresponding cache
entry and on next access the file will be re-read.

In order for caching to work, replaced g_key_file_free() with g_key_file_unref()
overall.