service: Connect ethernet networks when created
Ethernet networks are created once a cable is plugged in, i.e. carrier
is detected. According to the documentation in doc/service-api.txt,
ethernet networks will act as a Connect() D-Bus API call is issued with
the cable plug in.

Fix the code so that all ethernet services are connected at once when
created, provided that the network type is autoconnectable either by
default or by explicitely specifying the technology with the
DefaultAutoConnectTechnologies main.conf variable.

Reported by Mihai Neagu.