[connman] plugins: Dummy bluetooth plugin for rfkill setup
The bluetooth_legacy plugin only registers a bluetooth device when
bluetoothd has started up and indicated that an adapter is present;
and only at that time will connman set BT rfkill according to saved
settings.

This means that the intended behaviour of having connmand start before
bluetoothd and have rfkill set up before BT adapter is accessed does
not work. By implementing a dummy bluetooth driver we can initialize
bluetooth technology in connman and set rfkill to desired value
actually before bluetoothd starts up.

As a prerequisite, connmand has to wait for a HCI adapter to be present
in the system; otherwise setting up rfkill for BT won't work. This is
achieved by implementing a service that waits for an adapter to appear
outside connmand.

Service checks hardware ID from /etc/hw-release to determine whether
the wait is needed on the platform it is running or not.

[connman] plugins: Dummy bluetooth plugin for rfkill setup