[connman] Mark associating network as available. Contributes to JB#30297.
It may happen that autoscan is triggered after a connection request
has been made. Starting a scan clears networks' availability flags; if
the connection request proceeds to associating state before new scan
results are at hand the associating network may be freed and
connection attempt aborted. This happens even if state change and scan
results are received from supplicant in the same D-Bus signal but
State property is before BSSs in the message.

To avoid connection failure, mark the associating network as available
when state change to associating happens.