[connman] Clear disconnecting flag only on callback. Contributes to JB#30297.
As wifi->disconnecting is set to true only when a method call to
supplicant is made, do not clear the flag on interface state change,
but only in the callback. Otherwise, if a network connection is
requested after state change but before disconnection callback gets
called, wifi pending network isn't set properly and connection
progress won't be properly processed.