wpad: Unref the intended service
Unreference the service that has been stored in the connman_wpad
structure when calling  __connman_wpad_stop() instead of the service
given in the function argument. This ensures that if two services
point to the same interface index, the one referenced by wpad code
is the one that gets unreferenced.