dnsproxy: dns_name_length() returns incorrect length
When testing dnsproxy of connman, it was noticed that when sending a
cached response, the TTL field and CLASS field might be incorrect.

When looking into this, it was found that the TTL value might be
written one or two characters earlier in the DNS answer.

When looking in the send_cached_response() response routine,
update_cached_ttl() is called, which uses dns_name_length to shift the
pointer for the name of DNS Question and the name of DNS answer. When
further looking in dns_name_length(), strlen() is used to determine
the length of the name. However, strlen() is excluding the terminating
NULL character '\0', so 1 should be added to strlen for the length of
the name. dns_name_length() is also used in reply_query_type, here 1
is added to dns_name_length().