device: Remove unused device disconnected variable
The value of the struct connman_device disconnected variable is never
used in the device code or queried from other parts of the code.

The unnecessary connman_device_get_disconnected() function can be
removed, and as the disconnected variable is never used, so can
the connman_device_set_disconnected() function.

Conflicts:
	connman/src/device.c