doc: Move man pages under automake
This change prettifies man pages with <SYSCONFDIR> and /var/lib so
that it will show the actual value connman was compiled with, rather
than just hinting at some magical compile time variable.

Add only the man page autoconf .in files to the dist tar ball so
that running ./configure with a different prefix generates the man
pages using the given prefix. As the autoconf .in is not created by
standard rules, ensure also that the target directory is created.

Conflicts:
	connman/Makefile.am