service: Clear the Error property
It is been seen that if the service state has transitioned to failure
there is no way for it to get it back to idle. This fix allows the
state to be transitioned back to idle as part of handling clear_property
handler for error event.

Refer to commit 251d95755dd144c8bd6d3e3bd5d6a47f891f938f which fixes the
documentation for transitioning out of failure state.