technology: Cancel the pending reply timer
Sometimes connman crashes generating response for set_powered
after the timeout (pending_reply timeout). But before the expiry
of timer if technology is removed, the previous technology pointer
is no longer valid and on timeout it crashes. This change cancels
the timer and frees the memory before freeing the technology pointer.