ofono: Move "network" property to context struct
The current commit moves the "network" property to the context
structure. Many modifications are needed to take into account
some network modifications using a context instead of a modem struct.

A new function is needed : "get_context_with_network". It returns
a context from the list according to a network.

Conflicts:
	connman/plugins/ofono.c