wifi: p2p: Remove peers when technology is disabled.
If Wi-Fi was disabled, either by removal of a Wi-Fi dongle, or through
the D-Bus API, the Wi-Fi technology would sometimes be cleaned up
before gsupplicant would be signalled by wpa_s that the Peers were
removed. This means that the discovered peers are not cleaned up
properly as there is no wifi instance that can handle the callback of
the signal. Thus the GetPeers() method call could return a list of
Peers while Wi-Fi was disabled or removed.

Rewrite the peer-handling to be done more like the networks handling so
that when the technology is disabled it's possible to loop over the
peers and free them.