service: Create helper function for setting split routing
When setting split routing, the service order must also be updated
to match the functionality provided by __connman_service_get_order().
A VPN with split routing keeps its order at zero like all other services,
but one without split routing is sorted with a order of 10. The
service list is kept sorted in other parts of the code so it need not
be done here.

Conflicts:
    connman/src/service.c