service: Store passphrase after successfull WPS provisioning
When WPS Provisioning finished successfully, the passphrase gotten
from the WPS Credentials is stored first in the network structure
in handle_wps_completion() and then on the service structure in
service_indicate_state(). The problems is that the service's
settings file is created/updated through service_save() before
updating the service structure with the new passphrase. Therefore,
the passphrase is not actually written in the file.

This patch moves the updating of the passphrase before storing the
information in the settings file.