technology: Allow raw key for tethering
Although the gsupplicant component allows to use a raw key when a new
network is being created through the function add_network_security_psk,
the technology component does not allow it because in order to set the
property TetheringPassphrase the string's length must be within the
range [8, 63], otherwise it will be taken as wrong value without check
if it is a raw key.

This patch uses Wi-Fi passphrase/password checking already implemented
in service.c, check_passphrase(), exporting it in src/connman.h. The
connman_service_security argument is hard-coded to
CONNMAN_SERVICE_SECURITY_PSK because it is the only Wi-Fi security type
currently supported for thethering.