wifi: Clean up connected flag after disconnecting
After disconnecting, wifi->connected value is not coherent with the
actual interface state and it does not allow to start auto-scan. This
only happens when we are disconnecting from previous successful WPS
association. It is because in interface_disconnect_result() in
supplicant.c we do not have to remove the network and then disconnect
callback (disconnect_callback() in wifi.c) is called faster and
before signal "PropertiesChanged" is received which indicates "State"
changed to "disconnected". Therefore, when this signal is handled in
wifi.c:interface_state() the wifi->network was already removed by
disconnect_callback() and then interface_state() will intermediately
return without reseting wifi->connected flag to false. The other flags
like network->connected and wifi->disconnecting are correctly reseted
in wifi.c:disconnect_callback() but wifi->connected is missing.

This is a consequence of the fix done in
c82fe8682b742005086fd28b8681521f09747e9f because before it
disconnect_callback() in wifi.c was not even called.

(cherry picked from commit 87817b0ddfc9173c41ea6af4682be880f64eab4d)