nftables: Mark packets in the route table
The networking stack does the rerouting in the route table. There is a
check which prevents calling ip_reroute_me_harder when it was not
modified in route output chain. By moving our marking rule to the route
output chain the rerouting path is taken and we send our packets via the
right interface (which is defined in the policy routing table).

(cherry picked from commit 2a496a0ee5fb6510c0bb204861abd6bb87a359aa)