nftables: Drop connection tracking
We don't use the connection tracking for per session statistics, so just
drop it.

When we started with the Session API the goal was to support routing and
statistics per application. So far we implement the routing via policy
routing tables but never finished on the statistics side. Also the D-Bus
API does not expose counters for TX/RX.

It is quite likely we see eBPF support for cgroups in Linux
soon. Creating simple counters via eBPF and do some form of enforcement
in an extra component seems a better approach than stuffing all of this
into ConnMan. So for the time beeing remove the unused bits.

(cherry picked from commit d945ddd3fcc3311d64fbb6e81e7086c7f46a3867)