service: Update nameservers automatically
Automatically update nameserver information when they are appended
or removed to the resolver code so that nameservers for IPv6 can be
signalled after IPv4 has moved the service to state 'ready'. Create
a zero second timeout so that nameservers can be appended or removed
in a loop one by one with only one D-Bus PropertyChanged signal
being sent.

Verify that the service is either connected or the nameservers have
been removed when the service is inactive before sending the
PropertyChanged signal.