wifi: Propagete disconnect reason code to service
Currently, if the WiFi plugin gets disconnected we ignore the reason
why that happened. Instead we just cleanup and let the corresponding
Service go to the IDLE state. If autoconnect than decides to reconnect
to that service it might just fail again, because the AP decided to
block us.

To prevent this busy loop we introduce a new error code blocked on the
network and service layer. So whenever the AP decided to kick this
client from the network we set the error code.

BTW, wpa_supplicant sents first the reason code before it tells
ConnMan the interface got disconnected.

Tested by Saurav Babu.

(cherry picked from commit 3bda07204b11158dd18322db6e64dc92bdd07124)