session: Fix default route for ppp services
When a session is connected to a ppp service, ipconfig
will not contain a gateway IP address. Currently setting up
of the session routing table fails, due to
__connman_inet_add_default_to_table returning an error.

In that scenario, INADDR_ANY needs to be used as a gateway.

(cherry picked from commit 632abde59c091e1595ad566e56a88709349f4c17)