[connman] Reset the service->ignore flag
... when the service is associated the first network.

As a side effect of some sort of a ref-counting bug, the
service may not be freed after it's been removed from the
network (because something is holding the reference) and
then later re-associated with the new network with the same
name. In that case service->ignore would be true (because
it was set by __connman_service_remove_from_network) so
update_from_network needs to reset the ignore flag back
to false.