[connman] Added sailfish_wifi plugin. JB#37468
Replaces the original wifi plugin, as it was becoming unmaintainable.
Significant portions of the original one have been rewritten which
made applying upstream patches increasingly difficult. Also, upstream
is pushing their own replacement for wpa_supplicant and is likely to
stop supporting this plugin anyway.

This also removes support for signal_poll plugins from the connman core.
It was never merged upstream (for unknown reasons) and since nothing but
sailfish_wifi plugin is using this functionality, there's no need to have
it as a plugin.

[connman] WPS push button configuration. Fixes JB#37812