[connman] Added net.connman.Service.RestrictedPropertyChanged signal. JB#37923
Added GetProperty method and RestrictedPropertyChanged signal
to the net.connman.Service interface. The signal is emitted
instead of the PropertyChanged signal for the properties that
are not wide open to everyone. The client has to call GetProperty
to query the new value and to have its access checked.

Currently GetProperty only works for restricted properties in
order to reduce the amount of patching even though it might be
logical if it worked for any property. However, for a normal
wide-open property it will return net.connman.InvalidArguments
error.

The only use for GetProperty is to query the new value after
receiving the RestrictedPropertyChanged signal - the normal
GetProperties call return all the properties that the client
has access to.