[connman] Hidden networks require active scan. Fixes JB#38006
Once we find a hidden network, we actively scan for all known hidden
networks hoping that it's one of them. This results in two BSSes with
the same BSSID, one having SSID and one not. Both get associated with
the same network.

As long as there are hidden networks in range, active scans are
periodically repeated.