[connman] Slight refactoring of the D-Bus access control framework
Internal functions are moved to connman.h and prefixed with __
so that they don't get exported (and therefore become available
to the external plugins)

Removed access.h from the devel package because there are no
plans (or reasons) to develop external access control plugins.
Should that becomes necessary, this file can be moved back to
the public area. The same thing happened to wakeup_timer.h for
the same reasons.

Removed references to glib from access.h for consistency with
other plugin interfaces (no idea why they avoid glib.h in the
interface files but let's follow the pattern)