session: Add ContextIdentifier
The calling process, that implements the session API is identified by
the user ID as it is runs. All processes of the same user share the
same list of allowed bearers, and the same priority for choosing
between available bearers.

This extension allows processes to select a service context dependent
behaviour for which the routing decision is made.

This is an extention to session API interface. Helps to enable a
service differentiation for processes run by the same user. Allows
ConnMan to differentiate between bearer usage permissions and the
respective priorities based on the requested service type.

(cherry picked from commit 848f7820aa609185368a229fa729d67275988ada)