device: Fix enable/disable return code
The powered_pending status is an indication that the powered status is
in the process of being changed; the correct return code here should be
"in progress".

Without this patch, setting a technology "Powered" for a slow ofono
device in the process of powering up results in an immediate DBus
response even though the device is far from being ready.

With this patch, setting a technology "Powered" returns first when the
property has actually changed which is the desired behaviour.