iwd: Create wpa_s identifiers
The identifier has no special meaning except it needs to be unique. It
will also be used to save and load configs. We should use the same
identifier as we have created with wpa_supplicant. This is important
when the user switches from a wpa_supplicant setup to iwd one. This
avoids reconfigurering the system.

(cherry picked from commit f09d628f47900a6ee638792d7ded6c359cf8c3c1)