[connman] Use stored settings when creating a service from provider. JB#41684
This fix loads the settings from the service settings file in ConnMan
storage when a service is to be created from provider data.

Solves the issue of VPN autoconnect not working after connman-vpnd
is restarted or automatically started after crash.

The struct connman_service is zeroed at connman_service_create() and
setting of modification time is only required if there is no
settings file for the service.