[connman-vpn] Prevent connecting VPN if not online. Fixes JB#41705
This commit fixes the following:
 - Do not allow VPNs to be connected if there is no network
 - Do not leave VPNs connected when networks are disconnected one after
   another very quickly leaving VPN connection initialized due to
   default service change
 - Force VPNs to go through the default service always

This adds ConnMan state listener to vpnd. It first calls GetProperties
method of net.connman.Manager interface to get the state. If state is
not received vpnd does not accept VPNs to be connected until a signal
PropertyChanged with the ConnMan State has been received and the value
of the State is either "ready" or "online".

This adds ConnMan service listener to vpnd. When ConnMan is disconnected
the state is set to offline. When ConnMan is connected (no delay) or
vpnd is started (500ms delay) a function is added to glib main event loop.

Connman state is tracked within vpnd with a boolean. Value "true" =
online/ready, value "false" = offline/idle.

This adds a feature to support delayed connecting of VPNs that may
happen in a situation where ConnMan status is not queried yet and
connection attempt for a VPN comes from the system or from a user. In
this case a timeout function is added to the main loop that runs with 1s
interval. There can be only one timeout function and it will not be
added if the state has been queried by vpnd. If the state is not
online/ready the timeout will keep trying. If VPN is already connected
or connecting the timeout function will be removed from main loop.

This also adds a feature to disconnect all VPNs if the default changes
to NULL (no network) or when the default service changes in order to all
VPNs to use the default connection at all times. The disconnection and
autoconnection of VPNs happens only if the service type is not VPN.