[connman-vpn] Set autoconnect on each connect request. Fixes JB#39282
Moved toggling of autoconnect from plugins/vpn.c that would be done only
after VPN is successfully connected to vpn-provider.c when connection of
VPN is requested.

This way the autoconnect is completely user toggled, even when connman
is offline.

The autoconnect still is disabled if the plugin reports a failure, to
avoid VPN plugin with invalid config to be in autoconnecting loop.