[connman] Disconnect only connected VPNs. Fixes JB#41911
As a side effect, dnsproxy_default_changed() is called with NULL when
disconnecting a disconnected VPN service. This results in all DNS
servers to be removed.

CONNMAN_SERVICE_STATE_ASSOCIATION seems to be the final state of the
service, and CONNMAN_SERVICE_STATE_CONFIGURATION is a pre-state for
CONNMAN_SERVICE_STATE_READY so for these states it should be safe to
call disconnect in case of VPN.