[connman-vpn] Proper adding of default route for VPN. Fixes JB#41763
The route needs to be added with proper index (from provider, whose
index matches to vpn adapter created for it), and as gateway being NULL
so host entry is added to routing table. Supports both IPv4 and IPv6.

Added a function to check if IPv4/IPv6 address is unspecified (any
address 0.0.0.0 / ::/0).