[connman-vpn] Add default route option for VPN. Fixes JB#41902
This commit enables to define a "DefaultRoute" string for a provider in
VPN plugin that tells ConnMan not to use the VPN as default route. In
VPN plugin this can be set with vpn_provider_set_string(), value "false"
means that the VPN is not used as default route with ConnMan
connections.

The "DefaultRoute" string is sent by the VPN property_changed() to
ConnMan if the value for the string is set. If string has no value
nothing will be sent.

This string is handled by ConnMan as configuration strings
and src/provider.c has a function to check if the provider is being set
as default route or not. By default, it always reports true if the
string is not set.

If the VPN service having "DefaultRoute" as "false" is either attempted
to be switched as current_default or as default in service list the
service below the VPN service is set as current_default/default in
service list.

Additionally, moved is_any_addr() from vpn/vpn-provider.c to
src/ipaddress.c since it is used also elsewhere.