[wifi] Don't remove wifi networks too quickly. JB#41295
If wifi network is missing from the scan, give it a few more chances
to reappear and only when the number of recover attempts exceeds the
limit, notify connman core that the network is gone.

The recovery logic that was there before never really worked. This
one sometimes actually helps to avoid generating unnecessary D-Bus
traffic (i.e. removing and then adding back the same network).