[connman] Use DNS of depending non-default route VPN. Contributes to JB#42337
Added a functionality to the dnsproxy which includes the DNS servers of
the VPNs that depend on the current default service. The DNS servers of
a VPN are included only when it is not used as default route. The DNS
servers of a depending VPN service are enabled at creation stage and as
well as when default service changes.