[connman] Traverse service list only with debug. Contributes to JB#42337
Added check for debug mode when traversing service list for debugging
purposes. The list of services is traversed only when debug is set on
and service info needs to be printed.

Also, changed to use correct nameserver_add_all() function within
service.c instead of resolver.c one. The service's nameservers and
domains are added properly if the default has been changed.