[connman] Service autoconnect: use service type and reason. Contributes to JB#42337
Run auto connect for both transport services and VPN services if the
changed service is not a VPN service and reason is specified. If the
changed service is VPN run only VPN auto connect, which is triggered by,
for example, disconnection of a VPN that does not require service auto
connection to be run.

If the do_auto_connect() is run with VPN service without reason (the
reason is CONNMAN_SERVICE_CONNECT_REASON_NONE), ignore it.