[connman] Default service change checks simplified. Contributes to JB#42337
Simplified the checking of the current default in case it is a VPN. The
change is not to be made if the current default service depends on the
new default service and is connected. Otherwise, the VPN as current
default must be disconnected regardless of its state. The check of
whether the new service is allowed to be set is done earlier in
default_changed().

This fixes the issue of DNSs not being enabled as the change of default
service was prematurely terminated because only the depending service
pointer was checked for NULL. Occurs in a scenario when change is to be
made but the VPN is not connected and depends on the new default
service.