[vpn] Don't touch AutoConnect flag on VPN connection state changes. JB#42691
AutoConnect flag and connection state are pretty much independent things.
A service with AutoConnect flag off can be connected (e.g. manually), or
it could have AutoConnect on but be disconnected (if there are several
auto-connectable services of the same kind, only one of them actually gets
connected).

Change of AutoConnect flag may (albeit not necessarily) change the
connection state but not the other way around.

The last but not the least, AutoConnect flag is exposed via UI. Therefore,
it should only be turned on and off interactively by the user.