[connman] Restructure code from default change to sorting. JB#42337
This commit restructures the default_changed() default service change
checks and moves the usable content into the service_compare() function
to be used in service list sorting.

Most of the checks can be handled within service_compare(). Service
preference is now checked using the PreferredTechnologies list and used
in sorting.

This guarantees that the service on top is not changed if the new
service is ranked below it in the preferred services list.