[connman] Do not update preferred order if VPN is as default. JB#42337
This commit fixes the issue of letting a newly connected transport
service to overtake a VPN service being as default service. The typical
case is when VPN is connected and in CONNMAN_SERVICE_STATE_READY and a
new service, e.g., mobile data is enabled. Then mobile data would take
precedence for a moment, making the previous default, the VPN to
disconnect, and after that the service sorting takes place where the
transport service of the VPN is set as default and VPN is connected
again.