vpn: Reset flags before calling vpn_newlink() if VPN interface was up
This fixes the issue of VPN provider not being set ready as VPN
interface was already up when vpn_notify() was called. It is required to
reset the flags before calling vpn_newlink(), otherwise the state of the
VPN provider is not changed.

This would result in a situation where VPN provider settings were not
sent to connmand and later adding the routes of a VPN will fail as the
interface index was not updated. Also other information (interface type,
nameservers, domain) of the VPN provider is outdated. This happens when
same VPN is re-connected after transport service has been changed.