[vpn] Explicitely remove VPN service if provider removed. Fixes JB#42797
Remove VPN service from service.c service_list explicitely by using
__connman_service_remove() when provider is to be removed. In our fork
it is not enough to call unref() for the service as list has a weak
reference for each. Because of thid difference explicit removal is
required in order to remove the VPN completely instead of haunting in
the service_list.