[connman] Introduce IPv6 support for firewall. Contributes to JB#42674
This commit adds IPv6 support to firewall.c. Two new functions are added
to connman.h which allow to add and remove rules using IPv6
functionality that was added to iptables.c. This commit does not change
functionality of firewall.c, new functions are:
 - __connman_firewall_add_ipv6_rule()
 - __connman_firewall_remove_ipv6_rule()

The firewall functions operate on higher level than the iptables.c
functions so a clear separation of rule adding and removal is decided to
be implemented for IPv4 and IPv6. This abstracts the use of iptables and
for internal functionality this kind of separation of concerns here is
clarifying things instead of having to give a specific type for each
firewall function call.