[connman] Iptables restore, commit rules one by one. Fixes JB#43925
Change to commit individual rules one by one to reduce the probability
of crashes. Policies are set with iptc library functions so they are not
to be committed with __connman_iptables_commit().

This way each read rule is instantly restored to iptables instead of
adding all of them in a row and committing after last. Crash may occur
if something else is called via glib main that alters iptables between
each call to iptables_parse_rule().