[connman] Check that iptables table exists before flush. Fixes JB#43931
This commit introduces checking of the iptables table name from the
iptables table names file before flushing. It works for IPv4 and IPv6.
If table does not exist in the file, it is not flushed.

The files to check are "/proc/net/ip_tables_names" for IPv4 and
"/proc/net/ip6_tables_names" for IPv6. The tables that are flushed are
kept as the same.