[connman] Do not change iptables policy if firewall fails. JB#43998
This changes the firewall loading by setting first the rules for general
firewall and to enable default chain policies only if it succeeds. This
will prevent the device becoming a brick in sense of no networking in or
even out (as the DNS will be prevented) if a single rule in the configs
is wrong.