[connman] Support removal of firewall configs with reload. JB#44071
This adds support for removing the rules defined in firewall
configurations. Each rule has a record of the configuration file it was
loaded from and this change utilizes that to remove the rules from a
removed configuration files from all generic, tethering, dynamic and
currently used rules. The rules in generic, tethering and currently used
lists are removed from iptables before complete removal from the lists.

The configuration files that are stored when new configuration files are
loaded are used to check which files have been removed. This list is
cleaned up from the removed entries as well.