[connman] Apply new firewall rules in reload for old services. JB#44071
This commit adds applying of new firewall rules for old services that
did not previously have dynamic rules set and are online. The existing
service_state_changed() is used here as it contains state checking for
the service as well as it checks if the service has dynamic rules set.