iptables: Set ip6t_ip6 flags if IPv6 rule protocol is set.
Flags is required to be set as IP6T_F_PROTO if protocol for IPv6 rule is
being set (INCLUDEDIR/linux/netfilter_ipv6/ip6_tables.h). Without this
being set, ICMPv6 rules, for example, are installed to ip6tables but no
packet matches the rule as protocol check is skipped in kernel.