unit: Add unit test for iptables.c longjmp() use
This adds a unit test to run at build for testing longjmp() use in
iptables.c. The tests test both cases to 1) handle longjmp() when
setjmp() was properly called to save env and 2) exit() when longjmp() is
called via xtables exit_err() from a location that normally does not
call it.