[connman] Ignore not managed devices in eth plugin. JB#43928 JB#45035
Add listening of device status changes via notifier into ethernet
plugin. On some devices the ethernet plugin handles the developer mode
interface as it is defined to be as ethernet on configfs supporting
devices. When a device is set on and is a non-managed device it is
added to list which should not be accessed as it is controlled by an
other component (e.g., usb-moded). Only the index of the interface is
stored.

This fixes the issue of developer mode interface losing address and
connection after connman stop/restart.